17 skill composition (#21)
* [Add] Skills with json

#17 #14

+ 스킬 기본 정보 json으로 관리
+ 전투 필드 기본 로직 + json 구조 구안
~ 전투 씬 기존 코드 수정된 몬스터 구조에 맞춰서 수정

* [Add] Moving toward enemy

#14

+ 몬스터 가장 가까운 적 몬스터를 대상으로 이동

* [Feat] Normalizing speed of monster

#14

~ 몬스터 이동 속도 정규화
~ 차이 식별용 이동속도 다른 몬스터 정보 추가 구성

* [Add] Using skills

#17 #14

+ Normal 스킬 스프라이트
// 스킬 형태 별 조건 구성

* [Feat] Skill execution field

#14 #17

+ 스킬 형태 별 조건 구성
// 스킬 형태 별 시전 구성 필요

* [Add] Basic usage of skills

#17 #14

+ 기본적인 스킬 사용 로직 수립
+ 피격 시 넉백 기본 로직 수립

* [Feat] Modifying Battle

#14 #17

~ 테스트용 몬스터 추가
~ 넉백 랜덤 상수 추가
+ 몬스터 거리 유지 로직 추가

* [Feat] Case of possesing multiple skills

#17 #14

~ SkillStat 변수명 변경: skilType => skillType
~ Dash의 경우, 넉백 계수 더 높게 설정
~ 적 타겟팅 후 이동 방향 랜덤 상수 설정

* [Feat] Modifying camera

#14

~ 카메라 이동 범위 SerealizedField로 구성

* [Add] dead of monster

#17 #14

+ 몬스터 데미지 기본 계산 로직
+ 몬스터 죽음 상태 기본 로직

* [Feat] Floor damage

#14 #17

~ 데미지 내림 -> 정수로 맞춤

+++ SlimeGG/Assets/Scripts/Monsters/Objects/MonsterAbilityStat.cs
@@ -1,7 +1,7 @@
 [System.Serializable]
 public class MonsterAbilityStat : ElementStat
 {
-    public int hp { get; set; }
-    public int def { get; set; }
-    public int spd { get; set; }
+    public float hp { get; set; }
+    public float def { get; set; }
+    public float spd { get; set; }
 }
